crud cliente pet

+++ delete/deleteCliente.php
@@ -5,8 +5,23 @@ global $pdo;
 
 $idDeletar = $_GET['delete'];
 
+
+$sql = "SELECT pet.id FROM pet WHERE cliente_id = :id";
+
+$stmt = $pdo->prepare($sql);
+$stmt->execute(["id"=>$idDeletar]);
+$result = $stmt->fetch();
+
+
+
+if ($result != false) {
+    $_SESSION['erro'] = "Não é possivel deletar, o cliente tem um pet cadastrado";
+    header('Location: ../listar/listarCliente.php');
+    exit();
+}
 $PDOStatement = $pdo->query("DELETE FROM cliente where id=$idDeletar");
 $PDOStatement->execute();
 
 header('Location:../listar/listarCliente.php');
+
 ?>
